cpufreq: tuning interactive * cpufreq: interactive: remove hispeed_freq init restriction Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com> * CHROMIUM: cpufreq: interactive: calculate load before freq change
The update to cpu load for cpufreq rate changes was happening after the rate
change instead of before. This switches it to update the cpu load before the
cpufreq rate change so the old frequency is used for calculating the speed
adjusted cpu load. The old frequency was used for that active time, so it should
be used for calculating the speed adjusted cpu load.

BUG=chrome-os-partner:37673
TEST=power_LoadTest on Jerry
check we spend most of our time at the lower cpu frequencies

Change-Id: I2fff785561ad8aebd354ddd305e91d96c799a617
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/262925
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Signed-off-by: franciscofranco <franciscofranco.1990@gmail.com>

* drivers: cpufreq_interactive: handle error for module load fail

If the cpufreq_register_governor fails, resources for thread
speedchange_task should be released.
currently, concerned  resources are released in module_exit,
but if module loading fails, exit will not be called
and resources will remain acquired. this may leave kernel
in an unstable state.

Change-Id: Ic33f058c069d30bfd114fa1c1380325c8e00b51c
Signed-off-by: gaurav jindal <gauravjindal1104@gmail.com>
Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>

* interactive: avoid calling usecs_to_jiffies() repeatedly

Calling usecs_to_jiffies() repeatedly can be avoided if we store
timer_rate and timer_slack_val in jiffies from the first place.

Signed-off-by: Park Ju Hyung <qkrwngud825@gmail.com>
Signed-off-by: Alex Naidis <alex.naidis@linux.com>
Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>

Signed-off-by: sunilpaulmathew <sunil.kde@gmail.com>